Add info section
Added a section to form the information title and the four div containers with rounded borders and explanatory text. Also refactored the classes in order for CSS selectors to be more readable. Added a "section" prefix to every div class that separates the main sections of the webpage. The CSS file now uses descendant selectors to make the atttributes more neat. Next will do the quote section. Might want to revisit and see if any code can be condensed into single selector for multiple parts.